requirements.txt
numpy
python-telegram-bot==13.12
scikit-learn
pandas
